Updates to google_sign_in:
* Add a new `GoogleIdentity` interface, implemented by `GoogleSignInAccount`
* Move GoogleUserCircleAvatar to "widgets" library (exported by
  base library for backwards compatibility) and make it take an instance
  of `GoogleIdentity`, thus allowing it to be used by other packages that
  provide implementations of `GoogleIdentity`
